accident detection, toll assessment and notification, account balance reports

+++ src/main/scala/model/MinuteXWaySegDirKey.scala
@@ -3,7 +3,7 @@ package model
 /**
   * Created by root on 15.01.17.
   */
-case class MinuteXWaySegDirKey(minute: Int, xWay:Int, seg:Int, dir:Byte) {
+case class MinuteXWaySegDirKey(minute: Int, xWay:Int, seg:Byte, dir:Byte) {
 
   override def toString: String = "minute: " + minute + ", xWay: " + xWay + ", seg: " + seg + ", dir: " + dir
 }
