+++ src/main/scala/model/XwaySegDirKey.scala
@@ -3,7 +3,7 @@ package model
 /**
   * Created by root on 21.01.17.
   */
-case class XwaySegDirKey(xWay: Int, segment:Int, direction:Int) {
+case class XwaySegDirKey(xWay: Int, segment:Byte, direction:Int) {
 
   override def toString: String = s"($xWay , $segment , $direction)"
